mmer uint64

+++ headers/type.h
@@ -7,7 +7,7 @@
 using namespace std;
 
 typedef uint64_t icolor;
-typedef uint32_t mmer;
+typedef uint64_t mmer;
 typedef uint32_t iread;
 typedef uint32_t iposread;
 typedef uint64_t kmer;
